refactor: split vite plugin internals
Move Vite plugin discovery, virtual module generation, HMR helpers, dev middleware, path helpers, shared types, and virtual IDs into focused internal modules.

Keep src/vite.ts as the public orchestration entrypoint while preserving existing exports and runtime behavior for Issue #114.
